requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
PyJWT==2.8.0
pydantic==1.10.14
